Connect to database if previous attempts failed
If connecting to the database fails when application first starts up,
then subsequent database transactions will fail. This refactoring adds
a Database library module connection handle as a private variable that is
cached when the module is required. Callers can use getDb() to retrieve
the database connection handle and use it safely if there are no errors.
This is because the Database library will connect to the database if it
has not yet connected.